tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-direction_inherit.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-direction: inherit</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-direction-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="The computed value of 'animation-direction' is
   the same as the specified value">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes move {
         from {
            left: 8px;
         }
         to {
            left: 108px;
         }
      }
      p:first-child {
         margin-top: 124px;
      }
      div {
         width: 100px;
         height: 100px;
         background: red;
         position: fixed;
         top: 8px;
         left: 8px;
         animation-duration: 0.8s;
         animation-fill-mode: forwards;
         animation-iteration-count: 2;
      }
      div > div {
         background: blue;
         animation-direction: inherit;
      }
      .test {
         animation-direction: alternate;
      }
      .test, .test > div {
         animation-name: move;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var divs = document.querySelectorAll('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 2000);
         divs[1].addEventListener('animationend', function() {
            clearTimeout(timeout);
            var innerStyle = getComputedStyle(divs[1], null);
            var dir = innerStyle.getPropertyValue('animation-direction');
            var passed = (dir == 'alternate' && innerStyle.left == '8px');
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
         });
         divs[0].className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue square above should slide rightwards and then back again, and
   there should be no red.</p>
   <p>Testing...</p>
   <div>
      <div></div>
   </div>
</body>
</html>
